Store open caching logins in database
git-svn-id: https://svn.code.sf.net/p/geokretylog/code@106 ff536b33-f402-454f-b109-af6ef765b36a

+++ trunk/src/pl/nkg/geokrety/activities/AccountsActivity.java
@@ -159,7 +159,7 @@ public class AccountsActivity extends ManagedDialogsActivity {
 				.getStateHolder();
 
 		if (resultCode == RESULT_OK) {
-			int id = data.getIntExtra(User.ACCOUNT_ID,
+			long id = data.getLongExtra(User.ACCOUNT_ID,
 					ListView.INVALID_POSITION);
 			if (id == ListView.INVALID_POSITION) {
 				User account = new User(data.getExtras());
